% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrs.R
\name{calculate_slr}
\alias{calculate_slr}
\title{Calculate a Score-Based Likelihood Ratio}
\usage{
calculate_slr(
  sample1_path,
  sample2_path,
  project_dir = NULL,
  copy_samples = FALSE
)
}
\arguments{
\item{sample1_path}{File path to a handwriting sample saved in PNG file
format.}

\item{sample2_path}{File path to a second handwriting sample saved in PNG
file format.}

\item{project_dir}{Optional. A path to a directory where helper files will be
saved. If no project directory is specified, the helper files will be saved
to 'tempdir()' and deleted before the function terminates.}

\item{copy_samples}{TRUE or FALSE. If TRUE, the PNG files will be copied to
the project directory. If a project directory is not given, the samples
will not be copied.}
}
\value{
A number
}
\description{
Compares two handwriting samples scanned and saved a PNG images. The writing
in both samples is split into component shapes, or graphs, with
'handwriter::processDocument'. The graphs are grouped into clusters of
similar shapes with 'handwriter::get_clusterassignment'. The proportion of
graphs assigned to each cluster, called the cluster fill rates, are used as
writer profiles. The cluster fill rates are calculated with
'get_cluster_fill_rates'. A similarity score is calculated between the two
samples using a random forest trained with 'ranger'. The similarity score is
compared to reference distributions of 'same writer' and 'different writer'
similarity scores. The result is a score-based likelihood ratio that conveys
the strength of the evidence in favor of 'same writer' or 'different writer'.
}
\examples{
\dontrun{
# Compare two samples from the same writer
sample1 <- system.file(file.path("extdata", "w0030_s01_pWOZ_r01.png"), package = "handwriterRF")
sample2 <- system.file(file.path("extdata", "w0030_s01_pWOZ_r02.png"), package = "handwriterRF")
calculate_slr(sample1, sample2)

# Compare samples from two writers
sample1 <- system.file(file.path("extdata", "w0030_s01_pWOZ_r01.png"), package = "handwriterRF")
sample2 <- system.file(file.path("extdata", "w0238_s01_pWOZ_r02.png"), package = "handwriterRF")
calculate_slr(sample1, sample2)
}

}
