% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrs.R
\name{calculate_slr_with_clusters}
\alias{calculate_slr_with_clusters}
\title{Calculate a Score-Based Likelihood Ratio from Cluster Assignments}
\usage{
calculate_slr_with_clusters(
  sample1_clusters,
  sample2_clusters,
  rforest = random_forest,
  project_dir = NULL
)
}
\arguments{
\item{sample1_clusters}{A file path to cluster assignments created with
[handwriter::get_clusters_batch()]}

\item{sample2_clusters}{A file path to cluster assignments created with
[handwriter::get_clusters_batch()]}

\item{rforest}{Optional. A random forest trained with \pkg{ranger}. If
rforest is not given, the data object random_forest is used.}

\item{project_dir}{Optional. A path to a directory where helper files will be
saved. If no project directory is specified, the helper files will be saved
to tempdir() and deleted before the function terminates.}
}
\value{
A number great than or equal to zero
}
\description{
Calculates a score-based likelihood ratio between the cluster assignments of two
handwriting samples with these steps:
\enumerate{
    \item \code{\link[handwriter]{get_cluster_fill_counts}} counts the number of graphs assigned to each cluster.
    \item \code{\link{get_cluster_fill_rates}} calculates the proportion of graphs assigned to each cluster. The cluster fill rates serve as a writer profile.
    \item A similarity score is calculated between the cluster fill rates of the two documents using a random forest trained with \pkg{ranger}.
    \item The similarity score is compared to reference distributions of same writer and different
    writer similarity scores. The result is a score-based likelihood ratio that conveys the strength
    of the evidence in favor of same writer or different writer. For more details, see Madeline
    Johnson and Danica Ommen (2021) <doi:10.1002/sam.11566>.
}
}
\details{
This function is primarily useful for users who want to calculate score-based
likelihood ratios between large numbers of handwriting samples. Follow these
steps:
\enumerate{
   \item Run \code{\link[handwriter]{process_batch_dir}} on the folder containing the scanned handwriting samples.
   This splits the writing in the samples into component shapes, or graphs.
    \item Run \code{\link[handwriter]{get_clusters_batch}} on the output folder that contains the graphs. This groups the graphs from each sample
   into clusters of similar shapes.
   \item Run `calculate_slr_with_clusters()` on pairs of files output in the previous step.
}
}
\examples{
# Compare two samples from the same writer
c1 <- system.file(file.path("extdata", "clusters", "w0030_s01_pWOZ_r01.rds"),
                  package = "handwriterRF")
c2 <- system.file(file.path("extdata", "clusters", "w0030_s01_pWOZ_r02.rds"),
                  package = "handwriterRF")
calculate_slr_with_clusters(c1, c2)

# Compare samples from two writers
c2 <- system.file(file.path("extdata", "clusters", "w0030_s01_pWOZ_r01.rds"),
                  package = "handwriterRF")
c2 <- system.file(file.path("extdata", "clusters", "w0238_s01_pWOZ_r02.rds"),
                  package = "handwriterRF")
calculate_slr_with_clusters(c1, c2)

}
