% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{densities}
\alias{densities}
\title{Same Writer and Different Writer Densities}
\format{
A list with the following components:
\describe{
\item{same_writer}{A density of same writer scores created with the 'density' function.}
\item{diff_writer}{A density of different writer scores created with the 'density' function.}}
}
\usage{
densities
}
\description{
A list of 'same writer' and 'different writer' densities created with
'make_densities_from_rf' and the random forest 'rf'. A similarity score was
calculated for each distance value used to train 'rf'. The similarity score
is the proportion of decision trees that predicted 'same writer' for the
distance value. The 'same writer' density was created by applying the
'density' function to the 'same writer' similarity scores. Similarly, the
'different writer' density was created by applying the 'density' function to
the 'different writer' similarity scores.
}
\examples{
\dontrun{
  plot(densities$same_writer)
  plot(densities$diff_writer)
}

}
\keyword{datasets}
