% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-test.R
\name{downsample_diff_pairs}
\alias{downsample_diff_pairs}
\title{Downsample Pairs of Different Writer Distances}
\usage{
downsample_diff_pairs(df)
}
\arguments{
\item{df}{A data frame of distances}
}
\value{
A data frame
}
\description{
Downsample Pairs of Different Writer Distances
}
\examples{
dists <- get_distances(df = LND, distances = c("euclidean"))

}
