% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_docnames}
\alias{expand_docnames}
\title{Expand CSAFE Handwriting Database Document Names}
\usage{
expand_docnames(df, docname_col = "docname", suffix = "")
}
\arguments{
\item{df}{A data frame of containing at least one column of document names
from the CSAFE Handwriting Database.}

\item{docname_col}{The name of the column containing document names.}

\item{suffix}{A character string to add to the end of the new columns. Use ""
for no suffix.}
}
\value{
A data frame with new columns: writer, session, prompt, and repetition.
}
\description{
Documents from the CSAFE Handwriting Database contain the writer, session,
prompt, and repetition in the file name:
<writer>_<session>_<prompt>_<repetition>. Create writer, session, prompt, and
repetition columns by extracting these items from the document file names.
The original document name column will be kept.
}
\examples{
df <- expand_docnames(cfr, "docname", "")
df <- expand_docnames(cfr, "docname", "_1")

}
