% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_csafe_train_set}
\alias{get_csafe_train_set}
\title{Get Training Set}
\usage{
get_csafe_train_set(df, train_prompt_code)
}
\arguments{
\item{df}{A data frame of cluster fill rates created with
'get_cluster_fill_rates'}

\item{train_prompt_code}{Which prompt to use in the training set: "pLND", "pPHR", "pWOZ", or "pCMB"}
}
\value{
A data frame
}
\description{
Create a training set from a data frame of cluster fill rates created with
'get_cluster_fill_rates'.
}
\examples{
train <- get_csafe_train_set(df = cfr, train_prompt_code = "pCMB")

}
