% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_ref_scores}
\alias{get_ref_scores}
\title{Get Reference Scores}
\usage{
get_ref_scores(rforest, df)
}
\arguments{
\item{rforest}{A \pkg{ranger} random forest created with
\code{\link{train_rf}}.}

\item{df}{A data frame of cluster fill rates created with
\code{\link{get_cluster_fill_rates}} with an added writer ID column.}
}
\value{
A list of scores
}
\description{
Create reference scores of same writer and different writer scores from a
data frame of cluster fill rates.
}
\examples{
\donttest{
get_ref_scores(rforest = random_forest, df = validation)
}

}
