% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_single_method_distances}
\alias{get_single_method_distances}
\title{Get Distances with a Single Distance Measure}
\usage{
get_single_method_distances(
  df,
  distance_measure = "euc",
  dist_col_label = "dist"
)
}
\arguments{
\item{df}{A data frame of cluster fill rates created with
`get_cluster_fill_rates`}

\item{distance_measure}{A distance measure used by 'stats::dist'}

\item{dist_col_label}{A name for the output distance column}
}
\value{
A data frame of distances
}
\description{
Calculate distances using 'stats::dist' between all pairs of cluster fill
rates for a single distance measure.
}
\examples{
distances <- get_single_method_distances(df = cfr, distance_measure = "euclidean", dist_col_label = "dist")

}
