% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-test.R
\name{get_train_test_sets}
\alias{get_train_test_sets}
\title{Get Training and Testing Sets}
\usage{
get_train_test_sets(dists, train_n, downsample = TRUE)
}
\arguments{
\item{dists}{A data frame of distances created with 'get_distances'}

\item{train_n}{The number of writers to assign to the training set}

\item{downsample}{True or False. True downsamples the different writer
distances in the training set to equal the number of same writer distances.
The test set is not changed.}
}
\value{
A list of data frames
}
\description{
Create training and testing sets from a data frame of distances created with
'get_distances'.
}
\examples{
dists <- get_distances(df = cfr, distances = c("euclidean"))
sets <- get_train_test_sets(dists, 70)
}
