% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-test.R
\name{get_train_test_writers}
\alias{get_train_test_writers}
\title{Get Writers for Train and Test Sets}
\usage{
get_train_test_writers(dists, train_n)
}
\arguments{
\item{dists}{A data frame of distances created with `get_distances`}

\item{train_n}{The number of writers in 'dists' to assign to to the train set. The
remaining writers are assigned to the test set.}
}
\value{
A list
}
\description{
Randomly assign writers to train and test sets.
}
\examples{
dists <- get_distances(df = cfr, distances = c("euclidean"))
writers <- get_train_test_writers(dists, 70)

}
