% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handwriterRF-package.R
\docType{package}
\name{handwriterRF-package}
\alias{handwriterRF}
\alias{handwriterRF-package}
\title{handwriterRF: Handwriting Analysis with Random Forests}
\description{
Perform forensic handwriting analysis of two scanned handwritten documents. This package implements the statistical method described by Madeline Johnson and Danica Ommen (2021) \doi{10.1002/sam.11566}. Similarity measures and a random forest produce a score-based likelihood ratio that quantifies the strength of the evidence in favor of the documents being written by the same writer or different writers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CSAFE-ISU/handwriterRF}
  \item Report bugs at \url{https://github.com/CSAFE-ISU/handwriterRF/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Reinders \email{reinders.stephanie@gmail.com}

Authors:
\itemize{
  \item Iowa State University of Science and Technology on behalf of its Center for Statistics and Applications in Forensic Evidence [copyright holder, funder]
}

}
\keyword{internal}
