% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slrs.R
\name{interpret_slr}
\alias{interpret_slr}
\title{Interpret an SLR Value}
\usage{
interpret_slr(df)
}
\arguments{
\item{df}{A data frame created by \code{\link{calculate_slr}}.}
}
\value{
A string
}
\description{
Verbally interprent an SLR value.
}
\examples{
df <- data.frame("score" = 5, "slr" = 20)
interpret_slr(df)

df <- data.frame("score" = 0.12, "slr" = 0.5)
interpret_slr(df)

df <- data.frame("score" = 1, "slr" = 1)
interpret_slr(df)

df <- data.frame("score" = 0, "slr" = 0)
interpret_slr(df)

}
