% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histograms}
\alias{plot_histograms}
\title{Plot Histograms}
\usage{
plot_histograms(rforest, score = NULL)
}
\arguments{
\item{rforest}{A random forest created with [train_rf()]}

\item{score}{A similarity score calculated with [calculate_slr()]}
}
\value{
A ggplot2 plot of histograms
}
\description{
Plot histograms of same writer and different writers reference similarity
scores from a random forest created with [train_rf()]. Plot a vertical,
dashed line at a similarity score calculated with [calculate_slr()] to see
whether the score is more typical of the same writer or different writers
reference scores.
}
\examples{
plot_histograms(rforest = random_forest)

# Add a vertical line 0.1 on the horizontal axis.
plot_histograms(rforest = random_forest, score = 0.1)

}
