% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_ranger}
\alias{rf_ranger}
\title{Distances and a 'ranger' Random Forest}
\format{
A list with the following components:
\describe{
\item{rf}{A random forest created with 'randomForest'}
\item{dists}{A data frame of Euclidean distances between pairs of cluster fill
rates of documents used to train the random forest.}}
}
\usage{
rf_ranger
}
\description{
A list that contains a data frame of Euclidean distances between the cluster
fill rates of pairs of documents and a random forest trained on those
distances. The random forest was created with 'ranger'.
}
\examples{
# view the distances data frame
rf_ranger$dists

# view the random forest
rf_ranger$rf

}
\keyword{datasets}
