% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templateK40}
\alias{templateK40}
\title{Cluster Template with 40 Clusters}
\format{
A list containing the contents of the cluster template.
\describe{
\item{centers_seed}{An integer for the random number generator use to select the
starting cluster centers for the K-Means algorithm.}
\item{cluster}{A vector of cluster assignments
for each graph used to create the cluster template. The clusters are numbered sequentially 1, 2,...,K.}
\item{centers}{The final cluster centers produced by the K-Means algorithm.}
\item{K}{The number of clusters in the template.}
\item{n}{The number of training graphs to used to create the template.}
\item{docnames}{A vector that lists the training document from which each graph originated.}
\item{writers}{A vector that lists the writer of each graph.}
\item{iters}{The maximum number of iterations for the K-means
algorithm.}
\item{changes}{A vector of the number of graphs that
changed clusters on each iteration of the K-means algorithm.}
\item{outlierCutoff}{A vector of the outlier cutoff values calculated on
each iteration of the K-means algorithm.}
\item{stop_reason}{The reason the
K-means algorithm terminated.}
\item{wcd}{The within cluster
distances on the final iteration of the K-means algorithm. More specifically,
the distance between each graph and the center of the cluster to which it
was assigned  on each iteration. The output of \code{\link[handwriter]{make_clustering_template}}' stores
the within cluster distances on each iteration, but the previous iterations were removed here to reduce the file size.}
\item{wcss}{A vector of the
within-cluster sum of squares on each iteration of the K-means algorithm.}}
}
\usage{
templateK40
}
\description{
A cluster template created by \pkg{handwriter} with 40 clusters. This template
was created from 120 handwriting samples from the CSAFE Handwriting Database.
}
\details{
\pkg{handwriter} splits handwriting samples into component shapes called
graphs. The graphs are sorted into 40 clusters with a K-Means algorithm.
}
\examples{
# view number of clusters
templateK40$K

# view number of iterations
templateK40$iters

# view cluster centers
templateK40$centers

}
\keyword{cluster}
