% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_rf}
\alias{train_rf}
\title{Train a Random Forest}
\usage{
train_rf(
  df,
  ntrees,
  distance_measures,
  output_dir = NULL,
  run_number = 1,
  downsample_strata = c("match", "pLND", "pPHR", "pWOZ")
)
}
\arguments{
\item{df}{A data frame of cluster fill rates created with
'get_cluster_fill_rates'}

\item{ntrees}{An integer number of decision trees to use}

\item{distance_measures}{A vector of distance measures. Any combination of
"abs", "euc", "man", "max", and "cos" may be used.}

\item{output_dir}{A path to a directory where the random forest will be
saved.}

\item{run_number}{An integer used for both the set.seed function and to
distinguish between different runs on the same input data frame.}

\item{downsample_strata}{A character vector of strata for drawing a
stratified random sample of different writer pairs to equal the number of
same writer pairs. Use 'Null' to prevent downsampling and use all different
writer pairs. Use 'match' to treat all different writer pairs as a single
strata.}
}
\value{
A random forest
}
\description{
Train a random forest with 'ranger' from a data frame of cluster fill rates.
}
\examples{
train <- get_csafe_train_set(df = cfr, train_prompt_codes = "pCMB")
rforest <- train_rf(
  df = train,
  ntrees = 200,
  distance_measures = c("euc"),
  run_number = 1,
  downsample_strata = TRUE
)
}
