% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfr}
\alias{cfr}
\title{Cluster Fill Rates for 1200 CSAFE Handwriting Database Samples}
\format{
A data frame with 1200 rows and 42 variables:
\describe{
  \item{docname}{file name of the handwriting sample}
  \item{total_graphs}{The total number of graphs in the handwriting sample}
  \item{cluster1}{The number of graphs in cluster 1}
  \item{cluster2}{The number of graphs in cluster 2}
  \item{cluster3}{The number of graphs in cluster 3}
  \item{cluster4}{The number of graphs in cluster 4}
  \item{cluster5}{The number of graphs in cluster 5}
  \item{cluster6}{The number of graphs in cluster 6}
  \item{cluster7}{The number of graphs in cluster 7}
  \item{cluster8}{The number of graphs in cluster 8}
  \item{cluster9}{The number of graphs in cluster 9}
  \item{cluster10}{The number of graphs in cluster 10}
  \item{cluster11}{The number of graphs in cluster 11}
  \item{cluster12}{The number of graphs in cluster 12}
  \item{cluster13}{The number of graphs in cluster 13}
  \item{cluster14}{The number of graphs in cluster 14}
  \item{cluster15}{The number of graphs in cluster 15}
  \item{cluster16}{The number of graphs in cluster 16}
  \item{cluster17}{The number of graphs in cluster 17}
  \item{cluster18}{The number of graphs in cluster 18}
  \item{cluster19}{The number of graphs in cluster 19}
  \item{cluster20}{The number of graphs in cluster 20}
  \item{cluster21}{The number of graphs in cluster 21}
  \item{cluster22}{The number of graphs in cluster 22}
  \item{cluster23}{The number of graphs in cluster 23}
  \item{cluster24}{The number of graphs in cluster 24}
  \item{cluster25}{The number of graphs in cluster 25}
  \item{cluster26}{The number of graphs in cluster 26}
  \item{cluster27}{The number of graphs in cluster 27}
  \item{cluster28}{The number of graphs in cluster 28}
  \item{cluster29}{The number of graphs in cluster 29}
  \item{cluster30}{The number of graphs in cluster 30}
  \item{cluster31}{The number of graphs in cluster 31}
  \item{cluster32}{The number of graphs in cluster 32}
  \item{cluster33}{The number of graphs in cluster 33}
  \item{cluster34}{The number of graphs in cluster 34}
  \item{cluster35}{The number of graphs in cluster 35}
  \item{cluster36}{The number of graphs in cluster 36}
  \item{cluster37}{The number of graphs in cluster 37}
  \item{cluster38}{The number of graphs in cluster 38}
  \item{cluster39}{The number of graphs in cluster 39}
  \item{cluster40}{The number of graphs in cluster 40}
}
}
\source{
<https://forensicstats.org/handwritingdatabase/>
}
\usage{
cfr
}
\description{
A dataset containing cluster fill rates for for 1,200 handwriting samples
from the CSAFE Handwriting Database. The dataset was created by running
\code{\link{get_cluster_fill_rates}} on the cluster
fill counts data frame cfc. Cluster fill rates are the proportion of total
graphs assigned to each cluster.
}
\keyword{datasets}
