% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{densities}
\alias{densities}
\title{Same Writer and Different Writer Densities}
\format{
A list with the following components:
\describe{
\item{same_writer}{A density of same writer scores created with the 'density' function.}
\item{diff_writer}{A density of different writer scores created with the 'density' function.}}
}
\usage{
densities
}
\description{
A list of 'same writer' and 'different writer' densities. The densities were
created by using the 'density' function and a Guassian kernel. The densities
were fit to the 'same' and 'different' scores in rf$rf$votes.
}
\examples{
\dontrun{
  plot(densities$same_writer)
  plot(densities$diff_writer)
}

}
\keyword{datasets}
