% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{get_cluster_fill_rates}
\alias{get_cluster_fill_rates}
\title{Get Cluster Fill Rates}
\usage{
get_cluster_fill_rates(df)
}
\arguments{
\item{df}{A dataframe of cluster fill rates created with
\code{\link[handwriter]{get_cluster_fill_counts}}.}
}
\value{
A dataframe of cluster fill rates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{get_cluster_fill_rates} is deprecated.
Use \code{\link[handwriter]{get_cluster_fill_rates}} instead.
}
\examples{
\dontrun{
rates <- get_cluster_fill_rates(df = cfc)
}

}
