% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{get_cluster_fill_rates}
\alias{get_cluster_fill_rates}
\title{Get Cluster Fill Rates}
\usage{
get_cluster_fill_rates(df)
}
\arguments{
\item{df}{A data frame of cluster fill rates created with
\code{\link[handwriter]{get_cluster_fill_counts}}.}
}
\value{
A data frame of cluster fill rates.
}
\description{
Calculate cluster fill rates from a data frame of cluster fill counts created
with \code{\link[handwriter]{get_cluster_fill_counts}}.
}
\examples{
rates <- get_cluster_fill_rates(df = cfc)

}
