% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_csafe_train_set}
\alias{get_csafe_train_set}
\title{Get Training Set}
\usage{
get_csafe_train_set(df, train_prompt_codes)
}
\arguments{
\item{df}{A data frame of cluster fill rates created with
\code{\link{get_cluster_fill_rates}}}

\item{train_prompt_codes}{A character vector of which prompt(s) to use in the
training set. Available prompts are 'pLND', 'pPHR', 'pWOZ', and 'pCMB'.}
}
\value{
A data frame
}
\description{
Create a training set from a data frame of cluster fill rates from the CSAFE
Handwriting Database.
}
\examples{
train <- get_csafe_train_set(df = cfr, train_prompt_codes = 'pCMB')

}
