% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{get_distances}
\alias{get_distances}
\title{Get Distances}
\usage{
get_distances(cfr, distance = "euclidean")
}
\arguments{
\item{cfr}{A data frame of cluster fill rates created with
`get_cluster_fill_rates`}

\item{distance}{A distance method used by 'stats::dist'}
}
\value{
A data frame of distances
}
\description{
Calculate distances using 'stats::dist' between all pairs of cluster fill
rates.
}
\examples{
distances <- get_distances(cfr, "euclidean")

}
