% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get_score}
\alias{get_score}
\title{Calculate a Similarity Score}
\usage{
get_score(d, random_forest)
}
\arguments{
\item{d}{A data frame of distance(s) between two handwriting samples,
calculated with 'get_distances'. The distance(s) needs to be the
distance(s) used to train the random forest.}

\item{random_forest}{A 'ranger' random forest created with `train_rf`.}
}
\value{
A number
}
\description{
Use a trained random forest to produce a similarity score for the distance
between two handwriting samples.
}
\examples{
# Select the cluster fill rates for two handwriting samples
# Calculate the Euclidean distance between the cluster fill rates
d <- get_distances(cfr[1:2, ], c("abs", "euc"))
score <- get_score(d = d, random_forest = rf)

}
