% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest.R
\name{get_score}
\alias{get_score}
\title{Get Similarity Score}
\usage{
get_score(rf, d)
}
\arguments{
\item{rf}{A random forest created with 'randomForest'}

\item{d}{A distance between two handwriting samples, calculated with
'get_distances'}
}
\value{
A number
}
\description{
Use a trained random forest to produce a similarity score for the distance
between two handwriting samples.
}
\examples{
# Select the cluster fill rates for two handwriting samples
# Calculate the Euclidean distance between the cluster fill rates
d1 <- get_distances(cfr[1:2, ], c("abs", "euc"))
score <- get_score(rf$rf, d1)

d2 <- get_distances(cfr[1:3, ], c("abs", "euc"))
score <- get_score(rf$rf, d2)
}
