% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_forest}
\alias{random_forest}
\title{A \pkg{ranger} Random Forest and Data Frame of Distances}
\format{
A list with the following components:
\describe{
\item{rf}{A random forest created with \pkg{ranger} with settings:
importance = 'permutation', scale.permutation.importance = TRUE, and num.trees = 200.}
\item{distance_measures}{A vector of the distance measures used to train the random forest: c('abs', 'euc')}
}
}
\usage{
random_forest
}
\description{
A list that contains a trained random forest created with \pkg{ranger} and
the dataframe of distances used to train the random forest.
}
\examples{
# view the random forest
random_forest$rf

# view the distance measures used to train the random forest
random_forest$distance_measures

}
\keyword{datasets}
