% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_forest}
\alias{random_forest}
\title{A \pkg{ranger} Random Forest and Data Frame of Distances}
\format{
A list with the following components:
\describe{
#' \item{rf}{A random forest created with \pkg{ranger} with settings:
importance = 'permutation', scale.permutation.importance = TRUE, and num.trees = 200.}
\item{dists}{The data frame used to train the random forest. The data frame has
400 rows. Each row contains the absolute and Euclidean distances between the
cluster fill rates of two handwriting samples. If both handwriting samples are
from the same writer, the class is same. If the handwriting samples are from
different writers, the class is different. There are 200 same distances and
200 different distances in the data frame. The class is recorded in the match column.}
}
}
\usage{
random_forest
}
\description{
A list that contains a trained random forest created with \pkg{ranger} and
the data frame of distances used to train the random forest.
}
\examples{
# view the random forest
random_forest$rf

# view the distances data frame
random_forest$dists

}
\keyword{datasets}
